requirements.txt
asgiref==3.9.0
autopep8==2.3.2
Django==5.2.4
numpy==2.3.1
pandas==2.3.1
psycopg==3.2.9
psycopg2==2.9.10
pycodestyle==2.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2

